# docs/swagger.yaml
basePath: /api
definitions:
...
  backend.CreateGymSubscriptionRequest:
    properties:
      auto_renew:
        example: false
        type: boolean
      notes:
        example: Owner subscribe paket Premium
        type: string
      payment_method:
        example: transfer
        type: string
      saas_plan_public_id:
        example: 00000000-0000-4000-8000-000000000102
        type: string
      start_date:
        example: "2026-06-16"
        type: string
      status:
        example: paid
        type: string
    required:
    - saas_plan_public_id
    type: object
...
  backend.CreateSubscriptionRequest:
    properties:
      member_public_id:
        example: member-public-id
        type: string
      membership_package_public_id:
        example: package-public-id
        type: string
      start_date:
        example: "2026-06-16"
        type: string
    required:
    - member_public_id
    - membership_package_public_id
    type: object
...
  backend.LoginRequest:
    properties:
      email:
        example: owner@gym.com
        type: string
      gym_public_id:
        example: 00000000-0000-4000-8000-000000000001
        type: string
      password:
        example: password
        type: string
    required:
    - email
    - password
    type: object
  backend.LogoutRequest:
    properties:
      refresh_token:
        example: refresh-token
        type: string
    required:
    - refresh_token
    type: object
  backend.ManualCheckinRequest:
    properties:
      member_public_id:
        example: member-public-id
        type: string
      notes:
        example: Manual check-in karena kamera bermasalah
        type: string
    required:
    - member_public_id
    type: object
...
  backend.RefreshTokenRequest:
    properties:
      refresh_token:
        example: refresh-token
        type: string
    required:
    - refresh_token
    type: object
  backend.RegisterOwnerRequest:
    properties:
      gym_address:
        example: Jl. Contoh No. 1
        type: string
      gym_email:
        example: admin@stronggym.com
        type: string
      gym_name:
        example: Strong Gym
        type: string
      gym_phone:
        example: "08123456789"
        type: string
      owner_email:
        example: owner@stronggym.com
        type: string
      owner_name:
        example: Budi Owner
        type: string
      owner_password:
        example: secret123
        minLength: 6
        type: string
    required:
    - gym_name
    - owner_email
    - owner_name
    - owner_password
    type: object
...
  backend.ScanCheckinRequest:
    properties:
      qr_token:
        example: qr-token-random
        type: string
    required:
    - qr_token
    type: object
...
info:
  contact: {}
  description: Gym Management SaaS backend API.
  title: Gym Management SaaS API
  version: "1.0"
paths:
  /v1/admin/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend.LoginRequest'
      produces:
      - application/json
      responses: {}
      summary: Login admin/staff
      tags:
      - Admin Auth
  /v1/admin/auth/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Logout payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend.LogoutRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Logout admin/staff
      tags:
      - Admin Auth
  /v1/admin/auth/profile:
    get:
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get authenticated profile
      tags:
      - Admin Auth
  /v1/admin/auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend.RefreshTokenRequest'
      produces:
      - application/json
      responses: {}
      summary: Refresh access token
      tags:
      - Admin Auth
  /v1/admin/checkins/manual:
    post:
      parameters:
      - description: Manual check-in payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend.ManualCheckinRequest'
      responses: {}
      security:
      - BearerAuth: []
      summary: Manual check-in
      tags:
      - Check-ins
  /v1/admin/checkins/scan:
    post:
      parameters:
      - description: QR scan payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend.ScanCheckinRequest'
      responses: {}
      security:
      - BearerAuth: []
      summary: Admin QR check-in
      tags:
      - Check-ins
...
  /v1/admin/gym/subscription:
    get:
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get current gym SaaS subscription
      tags:
      - SaaS Subscription
    post:
      consumes:
      - application/json
      parameters:
      - description: Gym subscription payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend.CreateGymSubscriptionRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Subscribe gym to a SaaS plan
      tags:
      - SaaS Subscription
...
  /v1/admin/saas-plans:
    get:
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: List SaaS subscription plans
      tags:
      - SaaS Subscription
  /v1/admin/subscriptions:
    get:
      responses: {}
      security:
      - BearerAuth: []
      summary: List subscriptions
      tags:
      - Subscriptions
    post:
      parameters:
      - description: Create subscription payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend.CreateSubscriptionRequest'
      responses: {}
      security:
      - BearerAuth: []
      summary: Create manual subscription
      tags:
      - Subscriptions
  /v1/admin/subscriptions/{subscription_public_id}:
    get:
      responses: {}
      security:
      - BearerAuth: []
      summary: Detail subscription
      tags:
      - Subscriptions
  /v1/admin/subscriptions/{subscription_public_id}/cancel:
    patch:
      responses: {}
      security:
      - BearerAuth: []
      summary: Cancel subscription
      tags:
      - Subscriptions
  /v1/admin/subscriptions/{subscription_public_id}/expire:
    patch:
      responses: {}
      security:
      - BearerAuth: []
      summary: Expire subscription
      tags:
      - Subscriptions
...
  /v1/auth/google/callback:
    get:
      parameters:
      - description: Google authorization code
        in: query
        name: code
        required: true
        type: string
      - description: OAuth state
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Google OAuth callback
      tags:
      - Public Auth
  /v1/auth/register-owner:
    post:
      consumes:
      - application/json
      parameters:
      - description: Register owner payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend.RegisterOwnerRequest'
      produces:
      - application/json
      responses: {}
      summary: Register new gym owner
      tags:
      - Public Auth
  /v1/member/checkins/scan:
    post:
      parameters:
      - description: Public QR scan payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend.ScanCheckinRequest'
      responses: {}
      summary: Public QR check-in
      tags:
      - Public Member
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer {access_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"